requirements.txt
torch==2.1.0+cu118
torchvision==0.16.0+cu118
torchaudio==2.1.0+cu118
Pillow>=9.0.0
tqdm>=4.65.0
pathlib